Fixed Travis build
Removed the version from dev_requirements.txt in order to let travis
build with older python version. Also uses OrderedDict for older python
version in unittests otherwise the json oject key order is not kept and
the comparaison is then not possible.
The test for the log.exception() had the path to the test environment
which differ between test environment. Therefore now only test the start
and end of the backtrace equality.

Don't uses the "make test" for running test but call directly nose2 command.
The make target creates a python virtualenv which don't work as expected
in Travis as it already uses virtualenv.