BGDIINF_SB-1721: Fixed flask attribute filter crash
When the flask request contained a malformed json data and the json was
part of the filter configuration, the filter crashed with a
HTTPException which coudl have broken the flask application. For example
if the Flask application simply ignored the payload (e.g. for GET
request) but configured to log the json payload when available.