Fixed `%(asctime)s` and other formatting style in JSON formatter
When using the JSON formatter with the `asctime`  (e.g. `format={"time": "asctime", "msg"; "message"}`)
the `asctime` was not formatted and the message were `{"time": "asctime", "msg": "Your log message"}`.
Also when using `%(asctime)s` with other `%` style formatting, the time was correctly printed but logger
crashed with `KeyError: 'asctime'`

This was due to the hack used to format the message, where we overwrite the style `_fmt` protected property.
Now we uses a new style object for every json attribute for the formatting, this way we don't have
to hack protected properties anymore.
Also we use the unparsed `fmt` string to determine if we need to add the `asctime` attribute to the record.
This solved the issue using `asctime` without the formatting style (e.g. `%(asctime)s)`