BGDIINF_SB-2520: Breaking changes in Flask Attribute - Default value
Now in the Flask Attribute filter the empty string default value for missing
attributes has been removed.

This was an issue for missing flask context with the JSON formatter. If you
used the JSON formatter to access a Flask request attribute sub key, like for
example a specific header:

```yaml
format:
  request:
    headers:
      Accept: flask_request_headers.Accept
```

This would have raise an exception if the flask request was missing, because
in the log record we would have the attribute `flask_request_headers` set to
the default value `''` which is not a dict.

So now to avoid such combinations issue, the Flask Attribute filter always
require to have the context. In case of context missing to avoid exception
you need to use the new LogRecordIgnoreMissing record factory.