BGDIINF_SB-2520: Removed string constant from JSON Formatter
 #breaking-change

The string constant in the JSON Formatter caused troubled when the user
wanted to have an extra parameter in his log only if the extra was present.
In the case were the extra parameter was not present, the formatter could not
differentiate if it was a missing extra parameter or a string constant and
was therefore added as constant. This could cause some types inconsistency in
the json log output, for example if some log would add an extra as dictionary,
when this dict would be missing in the log record the log output would not be
a dictionary but a string constant.

To avoid such issue we simply removed the string constant feature. This can
still be achieved using the ConstAttribute filter.

Now if an attribute is missing from the log record, it will raise a ValueError
like the standard python formatter. You can avoid such exception by using the
`ignore_missing=True` configuration, in this case the default value will be
a proper one; `''`, `[]` or `{}` if the attribute is accessed via dotted key notation
with trailing dot.