PB-511: Django request context in logs
Middleware and filter that can be combined to add django request fields to all
logs within the scope of the request.
Middleware adds request object to a thread local variable. Filter adds the
request from thread to log record. Existing json filter can be used to
decide which request fields should be added to the log.