PB-931 Fix the exc_text field handling in JSON formatter
Adding the exc_text field to the record before formatting a message.
This allows for formatting the message through the django logging
facility
Removing the addition of exc_text to the message directly. Instead adding
it to the record, so that it can be formatted through the django logging
facility